+++ Controls/_input/resources/Field.ts
@@ -15,11 +15,9 @@ import {
     split
 } from 'Controls/_input/Base/InputUtil';
 import {FixBugs} from '../FixBugs';
-
-// tslint:disable-next-line:ban-ts-ignore
-// @ts-ignore
 import * as template from 'wml!Controls/_input/resources/Field/Field';
-import {splitValueForPasting, calculateInputType, transliterateInput} from './Util';
+import {splitValueForPasting, calculateInputType} from './Util';
+import transliterate from 'Controls/_input/resources/Transliterate';
 import {delay as runDelayed} from 'Types/function';
 import {IText} from 'Controls/decorator';
 
@@ -423,7 +421,7 @@ class Field<Value, ModelOptions>
             event.nativeEvent.altKey && event.nativeEvent.keyCode === T_KEY_CODE ||
             event.nativeEvent.keyCode === PAUSE_KEY_CODE
         ) {
-            transliterateInput(this._model.value, this._getFieldSelection()).then((value) => {
+            transliterate(this._model.value, this._getFieldSelection()).then((value) => {
                 this._updateField(value, this._getFieldSelection());
                 this._updateModel({value});
                 this._notifyEvent('valueChanged');
