+++ Controls/input.ts
@@ -60,6 +60,7 @@ import hoursFormat from 'Controls/_input/InputCallback/hoursFormat';
 import lengthConstraint from 'Controls/_input/InputCallback/lengthConstraint';
 export {default as IECompatibleLineHeights} from 'Controls/_input/Area/IECompatibleLineHeights';
 export {default as MobileFocusController} from 'Controls/_input/resources/MobileFocusController';
+export {default as transliterate} from 'Controls/_input/resources/Transliterate';
 export {default as NewBaseViewModel} from './_input/BaseViewModel';
 export {default as AdapterMask} from 'Controls/_input/Adapter/Mask';
 export {default as isMaskFormatValid} from 'Controls/_input/Mask/isFormatValid';
